more reorganization and moving stuff around; not worth documenting all the little changes
